Fix a bug that manifested itself by returning an answer that was 3600 seconds too big right around the transition _to_ DST in Europe/London. Added a test for this.
The problem was that the logic in adding an hour to $zone_off to
handle repeated local times was bogus. To fix this, I got rid of the
extra hour, and instead, check to see if DST changed an hour ago. If
it did, and the change was to go back an hour, we subtract an hour
from the result.

Code tweaks so $zone_off in timelocal() is the real UTC offset, not
the offset * -1.